+++ tests/test_connection.py
@@ -1,12 +1,12 @@
 import json
 import os
+from pathlib import Path
 from tempfile import NamedTemporaryFile
 from textwrap import dedent
 from unittest import mock
 
 import pytest
-
-from tests.testing_utils.fixtures import TEST_DIR
+import tomlkit
 
 
 def test_new_connection_can_be_added(runner, snapshot):
@@ -643,3 +643,32 @@ def test_source_precedence(mock_connect, runner):
         "database": "database_from_connection_env",
         "role": "role_from_global_env",
     }
+
+
+def test_set_default_connection_fails_if_no_connection(runner):
+    with NamedTemporaryFile("w+", suffix=".toml") as tmp_file:
+        result = runner.invoke_with_config_file(
+            tmp_file.name, ["connection", "set-default", "foo"]
+        )
+
+    assert result.exit_code == 1
+    assert "Connection foo is not configured" in result.output
+
+
+def test_set_default_connection(runner):
+    def _change_connection(config_file, conn_name):
+        result = runner.invoke_with_config_file(
+            config_file.name, ["connection", "set-default", conn_name]
+        )
+        assert result.exit_code == 0, result.output
+        return tomlkit.loads(Path(tmp_file.name).read_text()).value
+
+    with NamedTemporaryFile("w+", suffix=".toml") as tmp_file:
+        tmp_file.write("[connections.conn1]\n[connections.conn2]")
+        tmp_file.flush()
+
+        config = _change_connection(tmp_file, "conn1")
+        assert config["default_connection_name"] == "conn1"
+
+        config = _change_connection(tmp_file, "conn2")
+        assert config["default_connection_name"] == "conn2"